timepicker refactor
