+++ lib/FocusLink/FocusLink.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import css from './FocusLink.css';
 import { getNextFocusable } from '../../util/getFocusableElements';
-import separateComponentProps from '../../util/separateComponentProps';
 
 const propTypes = {
   children: PropTypes.oneOfType([
@@ -12,19 +11,27 @@ const propTypes = {
   ]),
   className: PropTypes.string,
   component: PropTypes.string,
+  sendOnFocus: PropTypes.func,
   showOnFocus: PropTypes.bool,
   tabIndex: PropTypes.number,
   target: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
   targetNextAfter: PropTypes.element,
 };
 
-const FocusLink = (props) => {
-  // eslint-disable-next-line react/forbid-foreign-prop-types
-  const [{ target, component, tabIndex, targetNextAfter }, rest] = separateComponentProps(props, FocusLink.propTypes);
-
+const FocusLink = ({
+  children,
+  className,
+  component,
+  sendOnFocus,
+  showOnFocus,
+  tabIndex,
+  target,
+  targetNextAfter,
+  ...rest
+}) => {
   let link = null;
 
-  function focusTarget() {
+  const focusTarget = () => {
     if (typeof target === 'string') {
       let id = target;
 
@@ -37,47 +44,47 @@ const FocusLink = (props) => {
         tgt.focus();
       }
     } else if (typeof target === 'object') {
-      props.target.focus();
+      target.focus();
     }
-  }
+  };
 
-  function focusNext() {
+  const focusNext = () => {
     let nextFocusable;
-    if (props.targetNextAfter) {
+    if (targetNextAfter) {
       nextFocusable = getNextFocusable(targetNextAfter, false);
     } else {
       nextFocusable = getNextFocusable(link, false);
     }
     nextFocusable.focus();
-  }
+  };
 
-  function handleClick(e) {
+  const handleClick = (e) => {
     e.preventDefault();
-    if (props.target) {
+    if (target) {
       focusTarget();
     } else {
       focusNext();
     }
-  }
+  };
 
-  function handleKeyPress(e) {
+  const handleKeyPress = (e) => {
     e.preventDefault();
     if (e.key === 'Enter') {
-      if (props.target) {
+      if (target) {
         focusTarget();
       } else {
         focusNext();
       }
     }
-  }
+  };
 
-  function getClass() {
+  const getClass = () => {
     return classNames(
       css.focusLink,
-      { [`${css.showOnFocus}`]: props.showOnFocus },
-      props.className,
+      { [`${css.showOnFocus}`]: showOnFocus },
+      className,
     );
-  }
+  };
 
   if (component) {
     const Component = component;
@@ -90,9 +97,10 @@ const FocusLink = (props) => {
         onClick={handleClick}
         onKeyPress={handleKeyPress}
         className={getClass()}
+        onFocus={sendOnFocus && focusTarget}
         {...rest}
       >
-        { props.children }
+        { children }
       </Component>
     );
   }
@@ -105,9 +113,10 @@ const FocusLink = (props) => {
       onClick={handleClick}
       onKeyPress={handleKeyPress}
       className={getClass()}
+      onFocus={sendOnFocus && focusTarget}
       {...rest}
     >
-      { props.children }
+      { children }
     </div>
   );
 };
