+++ lib/Timepicker/TimeDropdown.js
@@ -2,27 +2,91 @@
 * Display picker UI for Timepicker.
 */
 
-import React from 'react';
+import React, { useState, useRef, useCallback } from 'react';
 import PropTypes from 'prop-types';
-import Moment from 'moment';
-import classNames from 'classnames';
-import { injectIntl, FormattedMessage } from 'react-intl';
-import { extendMoment } from 'moment-range';
+import dayjs from 'dayjs';
+import customParseFormat from 'dayjs/plugin/customParseFormat';
+import { FormattedMessage, useIntl } from 'react-intl';
 import Button from '../Button';
 import IconButton from '../IconButton';
 import Layout from '../Layout';
 import { Row, Col } from '../LayoutGrid';
 import TextField from '../TextField';
 import css from './TimeDropdown.css';
+import FocusLink from '../FocusLink';
 
-const moment = extendMoment(Moment);
+dayjs.extend(customParseFormat);
+
+const keepInRange = (value, min, max, loop) => {
+  if (value > max) {
+    return loop ? min : max;
+  } else if (value < min) {
+    return loop ? max : min;
+  }
+  return value;
+};
+
+const rangedIncrement = (value, amount, min, max, add) => {
+  let local = value;
+  if (add) {
+    local += amount;
+    keepInRange(local, min, max);
+  } else {
+    local -= amount;
+    keepInRange(local, min, max);
+  }
+  return local;
+};
+
+const padZero = (value) => {
+  return value < 10 ?
+    `0${value}` :
+    value.toString();
+};
+
+// examine the passed time format to determine 12 or 24 hour format.
+function deriveHoursFormat(fmt) {
+  const ampmRE = new RegExp(/A/);
+  if (ampmRE.test(fmt)) {
+    return '12';
+  }
+  return '24';
+}
+
+const getListofDayPeriods = (locale) => {
+  // Build array of time stamps for convenience.
+  const dateArray = [];
+  while (dateArray.length < 24) {
+    dateArray.push(new Date(2022, 3, 10, dateArray.length));
+  }
+
+  const options = { hour: 'numeric' };
+  const dpOptions = new Set();
+  const df = new Intl.DateTimeFormat(locale, options);
+
+  dateArray.forEach((d) => {
+    const dateFields = df.formatToParts(d);
+    dateFields.forEach((f) => {
+      if (f.type === 'dayPeriod') {
+        dpOptions.add(f.value);
+      }
+    });
+  });
+  return [...dpOptions] || null;
+};
+
+const deriveDefaultTimePeriod = (value, hoursFormat, timeFormat, dayPeriods) => {
+  if (hoursFormat === '24') return null;
+  if (value) return dayjs(value, timeFormat).format('A');
+  return dayPeriods[0];
+};
 
 const propTypes = {
   hoursFormat: PropTypes.oneOf(['12', '24']),
   id: PropTypes.string,
   intl: PropTypes.shape({
-    formatMessage: PropTypes.func.isRequired
-  }).isRequired,
+    formatMessage: PropTypes.func
+  }),
   locale: PropTypes.string,
   mainControl: PropTypes.object,
   minuteIncrement: PropTypes.number,
@@ -39,308 +103,106 @@ const propTypes = {
   timeFormat: PropTypes.string.isRequired,
 };
 
-const defaultProps = {
-  hoursFormat: '24',
-  locale: 'en',
-  minuteIncrement: 1,
-  onSetTime: () => null,
-};
+const TimeDropdown = ({
+  hoursFormat: hoursFormatProp = 24,
+  id,
+  intl: intlProp,
+  locale: localeProp = 'en',
+  mainControl,
+  minuteIncrement = 1,
+  onBlur,
+  onHide,
+  onSetTime,
+  rootRef,
+  selectedTime,
+  timeFormat,
+}) => {
+  const intlContext = useIntl();
+  const intl = intlProp || intlContext;
+  const hoursFormat = hoursFormatProp || deriveHoursFormat(timeFormat);
+  const hourMax = useRef(hoursFormat === '24' ? 23 : 12).current;
+  const hourMin = useRef(hoursFormat === '24' ? 0 : 1).current;
+  const dayPeriods = hoursFormat === '12' && getListofDayPeriods(intl.locale);
+  const [hour, setHour] = useState(selectedTime && dayjs(selectedTime, timeFormat).format('HH'));
+  const [minute, setMinute] = useState(selectedTime && dayjs(selectedTime, timeFormat).format('mm'));
+  const [period, setPeriod] = useState(deriveDefaultTimePeriod(selectedTime, hoursFormat, timeFormat, dayPeriods));
+  const hoursInput = useRef(null);
+  const confirmButton = useRef(null);
 
-class TimeDropdown extends React.Component {
-  constructor(props) {
-    super(props);
-
-    moment.locale(this.props.locale);
-
-    // handle existing value...
-    let initMoment;
-    if (typeof props.selectedTime === 'undefined' || props.selectedTime === '') { // handle blank or cleared time...
-      initMoment = moment();
-    } else {
-      initMoment = moment(props.selectedTime, props.timeFormat, true);
-    }
-
-    let initialPresentationFormat = 'hh';
-    if (this.deriveHoursFormat(props.timeFormat) === '24') {
-      initialPresentationFormat = 'HH';
-    }
-
-    this.state = {
-      hoursFormat: this.deriveHoursFormat(),
-      mo: initMoment,
-      hour: initMoment.format(initialPresentationFormat),
-      minute: initMoment.format('mm'),
-      period: initMoment.format('A'),
-    };
-
-    this.enterTime = this.enterTime.bind(this);
-    this.getDOMContainer = this.getDOMContainer.bind(this);
-    this.incrementTime = this.incrementTime.bind(this);
-    this.handleChangeperiod = this.handleChangePeriod.bind(this);
-    this.deriveHoursFormat = this.deriveHoursFormat.bind(this);
-    this.handleBlur = this.handleBlur.bind(this);
-    this.getRootClass = this.getRootClass.bind(this);
-    this.confirmTime = this.confirmTime.bind(this);
-    this.setTimeHandleKeyDown = this.setTimeHandleKeyDown.bind(this);
-    this.hoursHandleKeyDown = this.hoursHandleKeyDown.bind(this);
-    this.buildState = this.buildState.bind(this);
-  }
-
-  // eslint-disable-next-line camelcase, react/no-deprecated
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    // if timeFormat updates, update the state's hours format (12/24hr time)
-    if (nextProps.timeFormat !== this.props.timeFormat) {
-      this.setState({
-        hoursFormat: this.deriveHoursFormat(nextProps.timeFormat),
-      });
-      this.buildState(nextProps);
-    }
-
-    if (nextProps.selectedTime !== this.props.selectedTime) {
-      this.buildState(nextProps);
-    }
-  }
-
-  componentWillUnmount() {
-    // Re-focus the timepicker input when the dropdown closes
-    this.props.mainControl.focus();
-    if (this.blurTO) {
-      clearTimeout(this.blurTO);
-      this.blurTO = null;
-    }
-
-    if (this.hoursBlurTO) {
-      clearTimeout(this.hoursBlurTO);
-      this.hoursBlurTO = null;
-    }
-  }
-
-  buildState(props) {
-    let initMoment;
-    if (props.selectedTime === '') {
-      initMoment = moment();
-    } else if (moment(props.selectedTime, props.timeFormat).isValid()) {
-      initMoment = moment(props.selectedTime, props.timeFormat, true);
-    } else if (moment(props.selectedTime, 'hh', true).isValid()) {
-      initMoment = moment(props.selectedTime, 'hh', true);
-    } else if (moment(props.selectedTime, 'h', true).isValid()) {
-      initMoment = moment(props.selectedTime, 'h', true);
-    } else {
-      initMoment = this.state.mo;
-    }
-
-    let initialPresentationFormat = 'hh';
-    if (this.deriveHoursFormat(props.timeFormat) === '24') {
-      initialPresentationFormat = 'HH';
-    }
-
-    this.setState({
-      mo: initMoment,
-      hour: initMoment.format(initialPresentationFormat),
-      minute: initMoment.format('mm'),
-      period: initMoment.format('A'),
-    });
-  }
-
-  // examine the passed time format to determine 12 or 24 hour format.
-  deriveHoursFormat(fmt) {
-    let _fmt;
-    if (!fmt) {
-      _fmt = this.props.timeFormat;
-    } else {
-      _fmt = fmt;
-    }
-    const periodRE = new RegExp(/A/);
-    if (periodRE.test(_fmt)) {
-      return '12';
-    }
-    return '24';
-  }
-
-  enterTime(e, unit) {
+  const modifyTime = useCallback((amount, unit, add) => {
     if (unit === 'hour') {
-      let parsedHours = parseInt(e.target.value, 10);
-      if (e.target.value.length > 1) {
-        const twelveHour = (this.state.hoursFormat === '12');
-        if (twelveHour) {
-          if (parsedHours > 12) {
-            parsedHours = 12;
-          }
-          if (parsedHours < 1) {
-            parsedHours = 1;
-          }
-        } else {
-          if (parsedHours > 23) {
-            parsedHours = 23;
-          }
-          if (parsedHours < 0) {
-            parsedHours = 0;
-          }
-        }
-      }
-      this.setState({
-        hour: parsedHours,
+      setHour((curHour) => {
+        let newHour = parseInt(curHour, 10);
+        newHour = rangedIncrement(newHour, amount, hourMax, hourMin, add);
+        newHour = padZero(newHour);
+        return newHour;
+      });
+    } else if (unit === 'minute') {
+      setMinute((curMinute) => {
+        let newMinute = parseInt(curMinute, 10);
+        newMinute = rangedIncrement(newMinute, amount, 0, 59, add);
+        newMinute = padZero(newMinute);
+        return newMinute;
       });
     }
-    if (unit === 'minute') {
-      let considered = e.target.value;
-      if (e.target.value.length > 2) {
-        /* only accept first 2 characters
-        *  since maxlength doesn't work on number fields in chrome
-        */
-        considered = e.target.value.substring(0, 2);
-      }
+  }, [hourMax, hourMin]);
 
-      let parsedMinutes = parseInt(considered, 10);
-      if (parsedMinutes > 59) {
-        parsedMinutes = 59;
-      }
-      if (parsedMinutes < 0) {
-        parsedMinutes = 0;
-      }
-      this.setState({
-        minute: parsedMinutes,
-      });
+  const enterTime = (e, unit) => {
+    if (unit === 'hour') {
+      let parsedHour = parseInt(e.target.value, 10);
+      parsedHour = keepInRange(parsedHour, hourMin, hourMax);
+      setHour(parsedHour);
+    } else if (unit === 'minute') {
+      let parsedMinute = parseInt(e.target.value, 10);
+      parsedMinute = keepInRange(parsedMinute, 0, 59);
+      setMinute(parsedMinute);
     }
-  }
+  };
 
-  // format number with leading 0 if its less than 10...
-  handleBlur(e, unit) {
-    let value;
+  const handleBlur = (e, unit) => {
     const parsedValue = parseInt(e.target.value, 10);
-    if (parsedValue < 10) {
-      value = `0${parsedValue}`;
-      this.setState({
-        [unit]: value,
-      });
-    }
-  }
-
-  setTimeHandleKeyDown(e) {
-    if (e.keyCode === 9 && !e.shiftKey) { // tab
-      // refocus the datepicker textfield if it's tabbed out...
-      this.props.onBlur(() => {
-        this.blurTO = setTimeout(() => {
-          this.props.mainControl.focus();
-        }, 20);
-      });
-    }
-  }
-
-  hoursHandleKeyDown(e) {
-    if (e.keyCode === 9 && e.shiftKey) { // tab
-      // refocus the datepicker textfield if the users shift-tabs out...
-      this.props.onBlur(() => {
-        this.hoursBlurTO = setTimeout(() => {
-          this.props.mainControl.focus();
-        }, 20);
-      });
-    }
-  }
-
-  getDOMContainer() {
-    return this.props.rootRef.current;
-  }
-
-  incrementTime(increment, unit, add) {
+    padZero(parsedValue);
     if (unit === 'hour') {
-      let maxHours;
-      let minHours;
-      if (this.deriveHoursFormat() === '12') {
-        maxHours = 12;
-        minHours = 1;
-      } else {
-        maxHours = 23;
-        minHours = 0;
-      }
-      this.setState((curState) => {
-        const newState = Object.assign({}, curState);
-        let newHour = parseInt(newState.hour, 10);
-        if (add) {
-          newHour += increment;
-          if (newHour > maxHours) {
-            newHour = minHours;
-          }
-        } else {
-          newHour -= increment;
-          if (newHour < minHours) {
-            newHour = maxHours;
-          }
-        }
-
-        // take care of the leading 0...
-        if (newHour < 10) {
-          newState.hour = `0${newHour}`;
-        } else {
-          newState.hour = newHour.toString();
-        }
-
-        return newState;
-      });
+      setHour(parsedValue);
+    } else {
+      setMinute(parsedValue);
     }
+  };
 
-    if (unit === 'minute') {
-      this.setState((curState) => {
-        const newState = Object.assign({}, curState);
-        let newMinute = parseInt(newState.minute, 10);
-        if (add) {
-          newMinute += increment;
-          if (newMinute > 59) {
-            newMinute = 0;
-          }
-        } else {
-          newMinute -= increment;
-          if (newMinute < 0) {
-            newMinute = 59;
-          }
-        }
+  const handleChangePeriod = () => {
+    setPeriod(prev => { return prev === 'AM' ? 'PM' : 'AM'; });
+  };
 
-        if (newMinute < 10) {
-          newState.minute = `0${newMinute}`;
-        } else {
-          newState.minute = newMinute.toString();
-        }
+  const confirmTime = (e) => {
+    e.preventDefault();
+    if (onSetTime) onSetTime({ hour, minute, period });
+  };
 
-        return newState;
-      });
-    }
-  }
-
-  confirmTime() {
-    this.props.onSetTime(this.state);
-  }
-
-  handleChangePeriod() {
-    this.setState(prevState => ({ period: prevState.period === 'AM' ? 'PM' : 'AM' }));
-  }
-
-  getRootClass() {
-    return classNames(
-      css.timepickerContainer,
-    );
-  }
-
-  render() {
-    const hourMin = this.deriveHoursFormat() === '12' ? 1 : 0;
-    const hourMax = hourMin === 1 ? 12 : 23;
-    const { intl: { formatMessage } } = this.props;
-    return (
-      <div
-        role="form"
-        onBlur={this.handleBlurDropdown}
-        ref={this.props.rootRef}
-        className={this.getRootClass()}
-        id={`timeDD-${this.props.id}`}
-        data-test-timepicker-dropdown
-      >
-        <Row center={this.state.hoursFormat === '12' ? undefined : 'xs'}>
+  return (
+    <form
+      ref={rootRef}
+      className={css.timepickerContainer}
+      id={`timeDD-${id}`}
+      data-test-timepicker-dropdown
+      onSubmit={confirmTime}
+    >
+      <>
+        <FocusLink
+          sendOnFocus
+          target={confirmButton.current}
+        >
+          <span className="sr-only">
+            <FormattedMessage id="stripes-components.Timepicker.formStart" />
+          </span>
+        </FocusLink>
+        <Row center={hoursFormat === '12' ? undefined : 'xs'}>
           <Col xs={4} className={css.center}>
             <IconButton
               data-test-timepicker-dropdown-increment-minutes-button
               tabIndex="-1"
               icon="arrow-up"
-              id={`clickable-timeDD-${this.props.id}-next-hour`}
-              onClick={() => { this.incrementTime(1, 'hour', true); }}
+              id={`clickable-timeDD-${id}-next-hour`}
+              onClick={() => { modifyTime(1, 'hour', true); }}
             />
           </Col>
           <Col xs={1}>&nbsp;</Col>
@@ -349,72 +211,79 @@ class TimeDropdown extends React.Component {
               data-test-timepicker-dropdown-increment-hours-button
               tabIndex="-1"
               icon="arrow-up"
-              id={`clickable-timeDD-${this.props.id}-next-minute`}
-              onClick={() => { this.incrementTime(this.props.minuteIncrement, 'minute', true); }}
+              id={`clickable-timeDD-${id}-next-minute`}
+              onClick={() => { modifyTime(minuteIncrement, 'minute', true); }}
             />
           </Col>
         </Row>
-        <Row center={this.props.hoursFormat === '12' ? undefined : 'xs'}>
+        <Row center={hoursFormat === '12' ? undefined : 'xs'}>
           <Col xs={4}>
-            <TextField
-              data-test-timepicker-dropdown-hours-input
-              autoFocus
-              aria-label={formatMessage({ id: 'stripes-components.Timepicker.hoursLabel' })}
-              inputRef={(h) => { this.hourField = h; }}
-              placeholder="HH"
-              onKeyDown={this.hoursHandleKeyDown}
-              min={hourMin}
-              max={hourMax}
-              type="number"
-              value={this.state.hour}
-              onChange={(e) => { this.enterTime(e, 'hour'); }}
-              onBlur={(e) => { this.handleBlur(e, 'hour'); }}
-              id={`timeDD-${this.props.id}-hour-input`}
-              marginBottom0
-            />
+            <FormattedMessage id="stripes-components.Timepicker.hoursLabel">
+              {(ariaLabel) => (
+                <TextField
+                  data-test-timepicker-dropdown-hours-input
+                  autoFocus
+                  inputRef={hoursInput}
+                  aria-label={ariaLabel}
+                  placeholder="HH"
+                  min={hourMin}
+                  max={hourMax}
+                  type="number"
+                  value={hour}
+                  onChange={(e) => { enterTime(e, 'hour'); }}
+                  onBlur={(e) => { handleBlur(e, 'hour'); }}
+                  id={`timeDD-${id}-hour-input`}
+                  marginBottom0
+                  hasClearIcon={false}
+                />
+              )}
+            </FormattedMessage>
           </Col>
           <Col xs={1}>
             <Layout className={`display-flex flex-align-items-center ${css.colon}`}>:</Layout>
           </Col>
           <Col xs={4}>
-            <TextField
-              data-test-timepicker-dropdown-minutes-input
-              aria-label={formatMessage({ id: 'stripes-components.Timepicker.minutesLabel' })}
-              inputRef={(m) => { this.minuteField = m; }}
-              placeholder="MM"
-              type="number"
-              min="0"
-              max="59"
-              value={this.state.minute}
-              onChange={(e) => { this.enterTime(e, 'minute'); }}
-              onBlur={(e) => { this.handleBlur(e, 'minute'); }}
-              id={`timeDD-${this.props.id}-minute-input`}
-              marginBottom0
-            />
+            <FormattedMessage id="stripes-components.Timepicker.minutesLabel">
+              {(ariaLabel) => (
+                <TextField
+                  data-test-timepicker-dropdown-minutes-input
+                  aria-label={ariaLabel}
+                  placeholder="MM"
+                  type="number"
+                  min="0"
+                  max="59"
+                  value={minute}
+                  onChange={(e) => { enterTime(e, 'minute'); }}
+                  onBlur={(e) => { handleBlur(e, 'minute'); }}
+                  id={`timeDD-${id}-minute-input`}
+                  marginBottom0
+                  hasClearIcon={false}
+                />
+              )}
+            </FormattedMessage>
           </Col>
-          {this.state.hoursFormat === '12' &&
+          {hoursFormat === '12' &&
             <Col xs={3}>
               <Button
-                aria-label={formatMessage({ id: 'stripes-components.Timepicker.PeriodLabel' })}
-                buttonRef={(a) => { this.aField = a; }}
+                aria-label={intlContext.formatMessage({ id: 'stripes-components.Timepicker.PeriodLabel' })}
                 fullWidth
                 marginBottom0
-                onClick={this.handleChangeperiod}
-                id={`timeDD-${this.props.id}-period-toggle`}
+                onClick={handleChangePeriod}
+                id={`timeDD-${id}-period-toggle`}
               >
-                {this.state.period}
+                {period}
               </Button>
             </Col>
           }
         </Row>
-        <Row center={this.state.hoursFormat === '12' ? undefined : 'xs'}>
+        <Row center={hoursFormat === '12' ? undefined : 'xs'}>
           <Col xs={4} className={css.center}>
             <IconButton
               data-test-timepicker-dropdown-decrement-hours-button
               tabIndex="-1"
-              id={`clickable-timeDD-${this.props.id}-prev-hour`}
+              id={`clickable-timeDD-${id}-prev-hour`}
               icon="arrow-down"
-              onClick={() => { this.incrementTime(1, 'hour', false); }}
+              onClick={() => { modifyTime(1, 'hour', false); }}
             />
           </Col>
           <Col xs={1}>&nbsp;</Col>
@@ -422,39 +291,41 @@ class TimeDropdown extends React.Component {
             <IconButton
               data-test-timepicker-dropdown-decrement-minutes-button
               tabIndex="-1"
-              id={`clickable-timeDD-${this.props.id}-prev-minute`}
+              id={`clickable-timeDD-${id}-prev-minute`}
               icon="arrow-down"
-              onClick={() => { this.incrementTime(this.props.minuteIncrement, 'minute', false); }}
+              onClick={() => { modifyTime(minuteIncrement, 'minute', false); }}
             />
           </Col>
         </Row>
         <Row>
           <Col xs={12}>
             <Button
+              type="submit"
               data-test-timepicker-dropdown-confirm-button
               buttonStyle="primary"
               buttonClass={css.submitButton}
-              onKeyDown={this.setTimeHandleKeyDown}
               fullWidth
               marginBottom0
-              onClick={this.confirmTime}
-              id={`clickable-timeDD-${this.props.id}-set-time`}
+              ref={confirmButton}
+              id={`clickable-timeDD-${id}-set-time`}
             >
               <FormattedMessage id="stripes-components.Timepicker.SetTimeLabel" />
             </Button>
-            <input
-              className="sr-only"
-              aria-label={formatMessage({ id: 'stripes-components.Timepicker.LeaveDialog' })}
-              onFocus={this.props.onHide}
-            />
           </Col>
         </Row>
-      </div>
-    );
-  }
-}
+        <FocusLink
+          sendOnFocus
+          target={hoursInput.current}
+        >
+          <span className="sr-only">
+            <FormattedMessage id="stripes-components.Timepicker.formEnd" />
+          </span>
+        </FocusLink>
+      </>
+    </form>
+  );
+};
 
 TimeDropdown.propTypes = propTypes;
-TimeDropdown.defaultProps = defaultProps;
 
-export default injectIntl(TimeDropdown);
+export default TimeDropdown;
