+++ lib/Timepicker/Timepicker.js
@@ -1,245 +1,255 @@
-import React from 'react';
-import { FormattedMessage, injectIntl } from 'react-intl';
+import React, { useEffect, useRef, useState } from 'react';
+import { FormattedMessage, useIntl, intl as rawIntl } from 'react-intl';
 import PropTypes from 'prop-types';
 import { deprecated } from 'prop-types-extra';
-import moment from 'moment-timezone';
-import contains from 'dom-helpers/query/contains';
-import debounce from 'lodash/debounce';
-import uniqueId from 'lodash/uniqueId';
-import isEmpty from 'lodash/isEmpty';
-
-import RootCloseWrapper from '../../util/RootCloseWrapper';
-import Popper, { AVAILABLE_PLACEMENTS } from '../Popper';
-import IconButton from '../IconButton';
+import dayjs from 'dayjs';
+import localeData from 'dayjs/plugin/localeData';
+import utc from 'dayjs/plugin/utc';
+import timeZone from 'dayjs/plugin/timezone';
+import objectSupport from 'dayjs/plugin/objectSupport';
+import { uniqueId } from 'lodash';
+import { useDynamicLocale } from '../../hooks/useDynamicLocale';
+import FormField from '../FormField';
 import TextField from '../TextField';
+import SRStatus from '../SRStatus';
+import Popper, { AVAILABLE_PLACEMENTS } from '../Popper';
+import RootCloseWrapper from '../../util/RootCloseWrapper';
+import IconButton from '../IconButton';
 import TimeDropdown from './TimeDropdown';
-import { getLocalizedTimeFormatInfo } from '../../util/dateTimeUtils';
+import parseMeta from '../FormField/parseMeta';
+import nativeChangeFieldValue from '../../util/nativeChangeFieldValue';
+import { getLocaleDateFormat } from '../../util/dateTimeUtils';
+
+dayjs.extend(localeData);
+dayjs.extend(utc);
+dayjs.extend(timeZone);
+dayjs.extend(objectSupport);
+
+const twelveHourFormats = {
+  '0': '',
+  '1': 'h',
+  '2': 'h:',
+  '3': 'h:mm',
+  '4': 'h:mm',
+  '5': 'h:mm ',
+  '6': ['h:mm A', 'h:mm a'],
+  '7': ['h:mm A', 'h:mm a']
+};
+
+const containsUTCOffset = (value) => {
+  const offsetRegex = /T[\d.:]+[+-][\d]+$/;
+  const offsetRE2 = /T[\d:]+[-+][\d:]+\d{2}$/; // sans milliseconds
+  const offsetZRE = /\.\d{3}Z$/i;
+  return offsetRegex.test(value) || offsetRE2.test(value) || offsetZRE.test(value);
+};
+
+const defaultParser = (value, timezone, timeFormat, passThroughValue, intl) => {
+  if (!value) return '';
+  if (value === passThroughValue) return passThroughValue;
+  let time;
+  if (containsUTCOffset(value)) {
+    // if it has an offset, it probably came from the backend and
+    // can be expected to be utc (utc offset of 0) -
+    time = dayjs.utc(value, timeFormat);
+  } else {
+    time = dayjs(value, timeFormat);
+  }
+
+  // entered value that's fewer characters than the expected format will pass through..
+  // e.g. '2' in an expected full string of '2:35 AM'
+  return time.isValid() ? time.format(timeFormat) : value;
+};
+
+const defaultOutputFormatter = ({ value, timezone, passThroughValue, formats, intl }) => {
+  if (value === passThroughValue) return passThroughValue;
+  let isoTime = dayjs.utc(value, formats);
+  if (isoTime.isValid()) {
+    isoTime = isoTime.toISOString();
+    const timeSplit = isoTime.split('T');
+    return timeSplit[1];
+  }
+  return '';
+};
 
 const propTypes = {
   autoFocus: PropTypes.bool,
   disabled: PropTypes.bool,
   id: PropTypes.string,
-  input: PropTypes.object,
-  inputRef: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+  inputRef: PropTypes.oneOfType([PropTypes.shape({ current: PropTypes.element }), PropTypes.func]),
   intl: PropTypes.object,
   label: PropTypes.node,
   locale: PropTypes.string,
   marginBottom0: PropTypes.bool,
-  meta: PropTypes.object,
   modifiers: PropTypes.object,
   onChange: PropTypes.func,
+  outputFormatter: PropTypes.func,
+  parser: PropTypes.func,
   passThroughValue: deprecated(PropTypes.string, 'Use alternative design'),
   placement: PropTypes.oneOf(AVAILABLE_PLACEMENTS),
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
   screenReaderMessage: PropTypes.string,
   showTimepicker: PropTypes.bool,
+  timeFormat: PropTypes.string,
   timeZone: PropTypes.string,
+  useInput: PropTypes.bool,
   usePortal: PropTypes.bool,
   value: PropTypes.string,
 };
 
-const defaultProps = {
-  autoFocus: false,
-  modifiers: {},
-  placement: 'bottom',
-  screenReaderMessage: '',
-  marginBottom0: false,
-  usePortal: false,
-};
+const Timepicker = ({
+  inputRef,
+  intl: intlProp,
+  locale: localeProp,
+  modifiers,
+  onChange,
+  outputFormatter = defaultOutputFormatter,
+  passThroughValue,
+  parser = defaultParser,
+  placement,
+  screenReaderMessage,
+  showTimepicker,
+  timeFormat: timeFormatProp,
+  timeZone: timeZoneProp,
+  useInput,
+  usePortal,
+  value: valueProp,
+  ...inputProps
+}) => {
+  const input = useRef(null);
+  const hiddenInput = useRef(null);
+  const srStatus = useRef(null);
+  const container = useRef(null);
+  const dropdownRef = useRef(null);
+  const testId = useRef(uniqueId('-timepicker')).current;
+  const { localeLoaded } = useDynamicLocale();
+  const intlContext = useIntl();
+  const intl = intlProp || intlContext;
+  const [showDropdown, setShowDropdown] = useState(showTimepicker || false);
+  const [timeFormat] = useState(
+    timeFormatProp ||
+    getLocaleDateFormat({ intl, config: { hour: 'numeric', minute: '2-digit' } })
+  );
+  const [timePair, updateTimePair] = useState({
+    timeString: valueProp ? parser(
+      valueProp, // value
+      timeZoneProp || intl.timeZone, // timezone
+      timeFormat, // uiFormat
+      passThroughValue,
+      intl
+    ) : null,
+    formatted: valueProp ? outputFormatter({
+      value: valueProp,
+      timeZone: timeZoneProp || intl.timeZone,
+      passThroughValue,
+      formats: twelveHourFormats[valueProp.length]
+    }) : null
+  });
+  const [displayed, updateDisplayed] = useState(timePair.timeString);
+  const candidate = useRef(timePair);
 
-class Timepicker extends React.Component {
-  constructor(props) {
-    super(props);
+  let maybeUpdateValue;
 
-    if (typeof props.inputRef === 'function') {
-      this.textfieldRef = (ref) => {
-        props.inputRef(ref);
-        this.textfieldRef.current = ref;
+  // handle value changes form outside of the component via maybeUpdateValue...
+  // useEffect(() => {
+  //   if (input.current
+  //     && typeof valueProp !== 'undefined'
+  //     && valueProp !== timePair.timeString
+  //     && valueProp !== timePair.formatted) {
+  //     candidate.current = Object.assign(candidate.current, maybeUpdateValue(valueProp));
+  //     nativeChangeFieldValue(input, false, candidate.dateString);
+  //   }
+  // }, [valueProp, maybeUpdateValue, timePair.timeString, timePair.formatted]);
+
+  /*  maybeUpdateValue
+  *   runs on all changes to the main input and value props to parse
+  *   the supplied value and, if it's a valid time, call the onChange handler.
+  */
+  maybeUpdateValue = (value) => {
+    // handle blank values...
+    if (value === '') {
+      const blankTime = {
+        timeString: '',
+        formatted: ''
       };
-    } else {
-      this.textfieldRef = props.inputRef || React.createRef();
-    }
-    this.containerRef = React.createRef();
-    this.dropdownRef = React.createRef();
-    this.picker = null;
-    this.srSpace = null;
-
-    this.handleSetTime = this.handleSetTime.bind(this);
-    this.handleKeyDown = this.handleKeyDown.bind(this);
-    this.clearTime = this.clearTime.bind(this);
-    this.showTimepicker = this.showTimepicker.bind(this);
-    this.handleFieldClick = this.handleFieldClick.bind(this);
-    this.hideTimepicker = this.hideTimepicker.bind(this);
-    this.dbhideTimepicker = debounce(this.hideTimepicker, 10);
-    this.toggleTimepicker = this.toggleTimepicker.bind(this);
-    this.handleRootClose = this.handleRootClose.bind(this);
-    this.hideOnBlur = this.hideOnBlur.bind(this);
-    this.standardizeTime = this.standardizeTime.bind(this);
-    this.handleFieldChange = this.handleFieldChange.bind(this);
-    this.pickerIsFocused = this.pickerIsFocused.bind(this);
-    this.handlePickTime = this.handlePickTime.bind(this);
-    this.deriveHoursFormat = this.deriveHoursFormat.bind(this);
-    this.getPresentationValue = this.getPresentationValue.bind(this);
-    this.getLocalTime = this.getLocalTime.bind(this);
-
-    // Set time zone
-    this.timeZone = props.timeZone || props.intl.timeZone || Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-    // Set locale
-    this.locale = props.locale || props.intl.locale;
-    if (moment.locales().includes(this.locale)) {
-      moment.locale(this.locale);
-      this._timeFormat = moment.localeData()._longDateFormat.LT;
-    } else {
-      this._timeFormat = getLocalizedTimeFormatInfo(this.locale).timeFormat;
+      updateTimePair(blankTime);
+      updateDisplayed('');
+      return blankTime;
     }
 
-    this.formats = [moment.ISO_8601, this._timeFormat, 'YYYY-MM-DD LT', 'YYYY-MM-DD h:mm:ss A',
-      'YYYY-MM-DD HH:mm:ss', 'YYYY-MM-DD HH:mm', 'HH:mm:ssZ'];
-    // inputValue will eventually be rendered as the value of the TextField.
-    let inputValue = '';
-    let presentedValue = null;
-    // if we aren't using redux form...
-    if (typeof this.props.input !== 'undefined') {
-      if (this.props.input.value !== '') {
-        if (this.props.input.value === this.props.passThroughValue) {
-          presentedValue = this.props.passThroughValue;
-          inputValue = moment.tz(this.timeZone).format(this._timeFormat);
-        } else {
-          const dateTime = `${moment().format('YYYY-MM-DD')}T${this.props.input.value}`;
-          // handle case where input value could be utc time or ISO formatted datetime
-          inputValue = moment(this.props.input.value, this.formats, true).isValid() ?
-            moment.tz(this.props.input.value, this._timeFormat, this.timeZone).format(this._timeFormat) :
-            moment.tz(dateTime, this.timeZone).format(this._timeFormat);
-        }
+    let valueTimeObject;
+
+    if (containsUTCOffset(value)) {
+      valueTimeObject = dayjs(value);
+    } else {
+      // use strict mode to check validity  - incomplete dates, anything not conforming to the format will be invalid
+      valueTimeObject = dayjs(
+        value,
+        [timeFormat, twelveHourFormats[value.length]],
+        true
+      );
+    }
+
+    let timeValues;
+
+    // otherwise parse the value and update the datestring and the formatted date...
+    if (valueTimeObject.isValid()) {
+      const parsed = parser(
+        value, // value
+        timeZoneProp || intl.timeZone, // timezone
+        timeFormat, // uiFormat
+        passThroughValue,
+        intl,
+      );
+      if (parsed !== timePair.timeString) {
+        const hiddenValue = outputFormatter({
+          value,
+          timeZone: timeZoneProp || intl.timeZone,
+          passThroughValue,
+          formats: twelveHourFormats[value.length]
+        });
+        timeValues = { timeString: parsed, formatted: hiddenValue };
+        updateTimePair(current => {
+          const newTimePair = Object.assign(current, timeValues);
+          return newTimePair;
+        });
+        updateDisplayed(timeValues.timeString);
+        return timeValues;
       }
-    } else if (this.props.value === this.props.passThroughValue) {
-      presentedValue = this.props.passThroughValue;
-      inputValue = moment.tz(this.timeZone).format(this._timeFormat);
-    } else {
-      inputValue = moment(this.props.value, this._timeFormat, this.timeZone).format(this._timeFormat);
+      return {};
+    } else if (value !== timePair.timeString) { // if the date's not valid, we just update the datestring...
+      timeValues = {
+        timeString: value,
+      };
+      updateTimePair(current => {
+        const newTimePair = Object.assign(current, timeValues);
+        return newTimePair;
+      });
+      updateDisplayed(timeValues.timeString);
+      return timeValues;
     }
+    return {};
+  };
 
-    this.state = {
-      presentedValue,
-      timeString: inputValue,
-      showTimepicker: this.props.showTimepicker || false,
-      srMessage: '',
-
-      timeFormat: this._timeFormat,
-      timeZone: this.timeZone,
-
-      prevLocale: this.locale,
-      prevValue: undefined,
-      prevInputValue: undefined,
-    };
-
-    if (props.id) {
-      this.testId = props.id;
-    } else {
-      this.testId = uniqueId('timepick-');
+  const handleInputRef = (ref) => {
+    if (typeof inputRef === 'function') {
+      inputRef(ref);
+    } else if (inputRef) {
+      inputRef.current = ref;
     }
-  }
+    input.current = ref;
+  };
 
-  static getDerivedStateFromProps(props, state) {
-    const newState = {};
+  const toggleTimepicker = () => {
+    setShowDropdown(cur => !cur);
+  };
 
-    if (props.value !== state.prevValue) {
-      newState.prevValue = props.value;
-    }
+  const hideTimepicker = () => {
+    setShowDropdown(false);
+  };
 
-    if (props.input) { // if we're using redux-form....
-      // if value has changed....
-      if (props.input.value !== '' && props.input.value !== state.prevInputValue) {
-        if (props.input.value === props.passThroughValue) {
-          newState.timeString = moment.tz(state.timeZone).format(state.timeFormat);
-          newState.presentedValue = props.passThroughValue;
-        }
-        // if value is blank....
-      } else if (props.input.value === '' && props.input.value !== state.prevInputValue) {
-        newState.presentedValue = null;
-        newState.timeString = '';
-      }
-
-      newState.prevInputValue = props.input.value;
-    } else if (props.value !== '' && props.value !== state.prevValue) {
-      if (props.value === props.passThroughValue) {
-        newState.timeString = moment.tz(state.timeZone).format(state.timeFormat);
-        newState.presentedValue = props.passThroughValue;
-      } else {
-        newState.timeString = moment.tz(state.timeZone).format(state.timeFormat);
-      }
-    }
-
-    // adjust displayed time format for a change in locale
-    if (props.intl.locale !== state.prevLocale) {
-      moment.locale(props.intl.locale);
-
-      newState.prevLocale = props.intl.locale;
-      newState.presentedValue = null;
-
-      if (state.timeString !== '') {
-        newState.timeString = moment.tz(
-          state.timeString, state.timeFormat, state.timeZone
-        ).format(state.timeFormat);
-      }
-    }
-
-    return isEmpty(newState) ? null : newState;
-  }
-
-  // examine the passed time format to determine 12 or 24 hour format.
-  deriveHoursFormat(fmt) {
-    let _fmt;
-    if (!fmt) {
-      _fmt = this.state.timeFormat;
-    } else {
-      _fmt = fmt;
-    }
-    const ampmRE = new RegExp(/A/);
-    if (ampmRE.test(_fmt)) {
-      return '12';
-    }
-    return '24';
-  }
-
-  getLocalTime(t, f) {
-    if (t) {
-      return moment(t).format(f);
-    }
-    return moment().format(this._timeFormat);
-  }
-
-  handleFieldClick() {
-    this.toggleTimepicker();
-  }
-
-  handleRootClose() {
-    this.hideTimepicker();
-  }
-
-  showTimepicker() {
-    this.setState({
-      showTimepicker: true,
-    });
-  }
-
-  hideTimepicker(cb) {
-    this.setState({
-      showTimepicker: false,
-    }, cb);
-  }
-
-  toggleTimepicker() {
-    const current = this.state.showTimepicker;
-    this.setState({
-      showTimepicker: !current,
-    });
-  }
-
-  handleKeyDown(e) {
-    if (this.picker) {
+  const handleKeyDown = (e) => {
+    if (dropdownRef.current) {
       // const formattedDate = curDate.format(this._timeFormat);
       // let timeString;
       switch (e.keyCode) {
@@ -255,7 +265,7 @@ class Timepicker extends React.Component {
           break;
         case 27: // escape
           e.preventDefault();
-          this.hideTimepicker();
+          hideTimepicker();
           break;
         case 34: // pgDn
           e.preventDefault();
@@ -265,15 +275,15 @@ class Timepicker extends React.Component {
           break;
         case 13: // enter
           e.preventDefault();
-          if (this.state.showTimepicker) {
+          if (showDropdown) {
             // this.handleSetTime(e, curDate, formattedDate);
           } else {
-            this.showTimepicker();
+            setShowDropdown(true);
           }
           // timeString = `${formattedDate} selected`;
           break;
         case 9: // tab
-          this.hideTimepicker();
+          setShowDropdown(false);
           break;
         default:
       }
@@ -283,359 +293,149 @@ class Timepicker extends React.Component {
         case 13: // enter
         case 40: // down
           e.preventDefault();
-          this.showTimepicker();
+          setShowDropdown(true);
           break;
         default:
       }
     }
-  }
+  };
 
-  handleSetTime(e, hours, minutes, ampm) {
-    let standardizedTime;
-    if (e === null || e.type === 'click' || e.type === 'keydown') {
-      if (e) { e.preventDefault(); }
+  // const handleSetTime = (e, hours, minutes, period) => {
+  //   let standardizedTime;
+  //   if (e === null || e.type === 'click' || e.type === 'keydown') {
+  //     if (e) { e.preventDefault(); }
+  //     if (onChange) { onChange(e); }
+  //     input.current?.focus();
 
-      if (this.props.onChange) { this.props.onChange(e); }
-      this.textfieldRef.current.focus();
+  //     let tString;
+  //     if (deriveHoursFormat(timeFormat) === '12') {
+  //       tString = `${hours}:${minutes} ${period}`;
+  //     } else {
+  //       tString = `${hours}:${minutes}`;
+  //     }
 
-      let tString;
-      if (this.deriveHoursFormat() === '12') {
-        tString = `${hours}:${minutes} ${ampm}`;
-      } else {
-        tString = `${hours}:${minutes}`;
-      }
+  //     setTimeString(tString);
+  //     standardizedTime = standardizeTime(`${hours}:${minutes} ${period}`);
+  //     if (onChange) { onChange(e, standardizedTime); }
+  //   } else if (e.type === 'change') {
+  //     setTimeString(e.target.value);
+  //     standardizedTime = standardizeTime(`${hours}:${minutes}${period && (' ' + period)}`);
+  //     if (onChange) { onChange(e, standardizedTime); }
+  //   }
+  // };
 
-      this.setState({
-        presentedValue: null,
-        timeString: tString,
-      });
-
-      standardizedTime = this.standardizeTime(`${hours}:${minutes} ${ampm}`);
-      if (this.props.input) {
-        this.props.input.onChange(standardizedTime);
-        if (this.props.onChange) {
-          this.props.onChange(e, standardizedTime);
-        }
-      } else if (this.props.onChange) {
-        this.props.onChange(e, standardizedTime);
-      }
-    } else if (e.type === 'change') {
-      // if e.target.value is even a substring of props.passThroughValue, pass it through...
-
-
-      // be sure that value is parseable as a date in the required format...
-      // the boolean parameter suppresses moment's deprecation warning,
-      // preventing it from attempting a parse using independable js Date object.
-      this.setState({
-        timeString: e.target.value,
-      });
-
-      if (this.props.onChange) { this.props.onChange(e); }
-      if (this.props.input) {
-        if (moment(e.target.value, this._timeFormat, true).isValid()) {
-          this.setState({
-            timeString: e.target.value,
-          });
-          // convert date to ISO 8601-ish format for backend (everything after the 'T')
-          if (!ampm) {
-            standardizedTime = this.standardizeTime(`${hours}:${minutes}`);
-          } else {
-            standardizedTime = this.standardizeTime(`${hours}:${minutes} ${ampm}`);
-          }
-          // redux-form handlers take the value rather than the event...
-          this.props.input.onChange(standardizedTime);
-          if (this.props.onChange) {
-            this.props.onChange(e, standardizedTime);
-          }
-        }
-      }
+  // for vanilla react/non-final-form implementations that just get the input value.
+  const handleChange = (e) => {
+    candidate.current = Object.assign(candidate.current, maybeUpdateValue(e.target.value));
+    if (useInput && onChange) {
+      onChange(e, e.target.value, candidate.current.timeString);
+    } else if (typeof candidate.current.formatted === 'string' &&
+      candidate.current.formatted !== hiddenInput.current.value) {
+      nativeChangeFieldValue(hiddenInput, false, candidate.current.formatted);
     }
-  }
+  };
 
-  handleFieldChange(e) {
-    if (e.target.value === '') {
-      this.clearTime();
-      return;
+  // for final-form so it can have a native change event rather than a fabricated thing...
+  const onChangeFormatted = (e) => {
+    if (!useInput && onChange) {
+      const { timeString, formatted } = candidate.current;
+      onChange(e, formatted, timeString);
     }
-    const ptREx = new RegExp(`^${e.target.value}`, 'i');
-    if (ptREx.test(this.props.passThroughValue)) {
-      if (this.props.onChange) { this.props.onChange(e); }
-      if (this.props.input) {
-        this.props.input.onChange(e.target.value);
-      }
-      this.setState({
-        presentedValue: e.target.value,
-      });
-    } else {
-      let hoursMins = e.target.value.split(/[:\s]/g);
-      hoursMins = hoursMins.filter(val => val); // remove any blank items...
-      if (hoursMins.length > 0) {
-        let hours = hoursMins[0];
-        const intHours = parseInt(hours, 10);
-        if (intHours < 10) {
-          hours = `0${intHours}`;
-        }
+  };
 
-        if (hoursMins.length < 2) {
-          this.handleSetTime(e, hours, '00');
-          return;
-        }
-        let mins = hoursMins[1];
-        let ampm;
-        if (hoursMins.length < 3) {
-          if (mins.length > 2) { // handle 3:00PM (no space between minutes and meridiem)
-            if (this.deriveHoursFormat() === '12') {
-              ampm = 'PM';
-              if (/[aA]/.test(mins)) {
-                ampm = 'AM';
-              }
-              const intMins = parseInt(mins, 10);
-              if (intMins < 10) {
-                mins = `0${intMins}`;
-              } else {
-                mins = intMins.toString();
-              }
-              this.handleSetTime(e, hours, mins, ampm);
-            }
-          }
-          this.handleSetTime(e, hours, mins);
-          return;
-        }
-        ampm = hoursMins[2];
-        this.handleSetTime(e, hours, mins, ampm);
-      }
-    }
+  // const setFromCalendar = (value) => {
+  //   nativeChangeField(input, hideOnChoose, value);
+  //   if (hideOnChoose) {
+  //     setShowCalendar(false);
+  //   }
+  // };
+
+  const handlePickTime = ({ hour, minute, period }) => {
+    const displayTime = dayjs({ hour, minute, period }).format(timeFormat);
+    nativeChangeFieldValue(input, true, displayTime);
+    setShowDropdown(false);
+  };
+
+  const handleClearTime = () => {
+    updateTimePair({ timeString: '', formatted: '' });
+    nativeChangeFieldValue(input, true, '');
+  };
+
+  const portalElem = usePortal
+    ? document.getElementById('OverlayContainer')
+    : null;
+
+  const { readOnly, disabled, label } = inputProps;
+  const screenReaderFormat = timeFormat.split('').join(' ');
+
+  let ariaLabel;
+  if (readOnly || disabled) {
+    ariaLabel = `${label}`;
+  } else {
+    // eslint-disable-next-line
+    ariaLabel = intl.formatMessage({ id: 'stripes-components.Timepicker.ScreenReaderLabel' }, { label, screenReaderFormat, screenReaderMessage });
   }
 
-  clearTime(e) {
-    this.setState({
-      presentedValue: null,
-      timeString: '',
-    });
-    if (this.props.onChange) { this.props.onChange(e); }
-    if (this.props.input) {
-      this.props.input.onChange('');
-    }
-    this.textfieldRef.current.focus();
-  }
-
-  cleanForScreenReader(str) {
-    const newString = str.replace(/H/g, 'H ').replace(/m/g, 'm ').replace(/s/g, 's ');
-    return newString;
-  }
-
-  informScreenReader(str) {
-    this.setState({ srMessage: str });
-  }
-
-  pickerIsFocused() {
-    if (this.picker && contains(this.containerRef.current, document.activeElement) &&
-      document.activeElement !== document.body) {
-      return contains(this.picker.getDOMContainer(), document.activeElement);
-    }
-    return false;
-  }
-
-  hideOnBlur(e) {
-    if (this.pickerIsFocused()) {
-      if (this.props.input) {
-        this.props.input.onBlur(e);
-      }
-      this.hideTimepicker();
-      this.setState({ srMessage: '' });
-    }
-  }
-
-  standardizeTime(time) {
-    const isoTime = moment.tz(time, 'HH:mm A', this.timeZone).toISOString();
-    const timeSplit = isoTime.split('T');
-    return timeSplit[1];
-  }
-
-  handlePickTime(timeObject) {
-    this.handleSetTime(null, timeObject.hour, timeObject.minute, timeObject.period);
-    this.hideTimepicker();
-  }
-
-  getPresentationValue() {
-    if (this.state.presentedValue !== null) {
-      return this.state.presentedValue;
-    }
-    return this.state.timeString;
-  }
-
-  render() {
-    const {
-      disabled,
-      input,
-      intl,
-      label,
-      marginBottom0,
-      modifiers,
-      placement,
-      readOnly,
-      required,
-      screenReaderMessage,
-      usePortal,
-      ...rest
-    } = this.props;
-
-    const screenReaderFormat = this.cleanForScreenReader(this._timeFormat);
-
-    let ariaLabel;
-    if (readOnly || disabled) {
-      ariaLabel = `${label}`;
-    } else {
-      // eslint-disable-next-line
-      ariaLabel = intl.formatMessage({ id:'stripes-components.Timepicker.ScreenReaderLabel'} ,{ label, screenReaderFormat, screenReaderMessage });
-    }
-
-    let endElement;
-    if (this.state.timeString !== '' || this.state.presentedValue !== null) {
-      endElement = (
-        <>
-          <FormattedMessage id="stripes-components.clearFieldValue">
-            {([clearFieldValueAriaLabel]) => (
-              <IconButton
-                key="clearButton"
-                onClick={this.clearTime}
-                aria-label={clearFieldValueAriaLabel}
-                id={`timepicker-clear-button-${this.testId}`}
-                tabIndex="-1"
-                icon="times-circle-solid"
-              />
-            )}
-          </FormattedMessage>
-          <FormattedMessage id="stripes-components.showOrHideTimepicker">
-            {([showOrHideTimepickerAriaLabel]) => (
-              <IconButton
-                key="timepickerButton"
-                onClick={this.toggleTimepicker}
-                onKeyDown={this.handleKeyDown}
-                aria-label={showOrHideTimepickerAriaLabel}
-                id={`timepicker-toggle-button-${this.testId}`}
-                tabIndex="-1"
-                icon="clock"
-              />
-            )}
-          </FormattedMessage>
-        </>
-      );
-    } else {
-      endElement = (
-        <FormattedMessage id="stripes-components.showOrHideTimepicker">
-          {([showOrHideTimepickerAriaLabel]) => (
-            <IconButton
-              key="timepickerButton"
-              onClick={this.toggleTimepicker}
-              onKeyDown={this.handleKeyDown}
-              aria-label={showOrHideTimepickerAriaLabel}
-              id={`timepicker-toggle-button-${this.testId}`}
-              tabIndex="-1"
-              icon="clock"
-            />
-          )}
-        </FormattedMessage>
-      );
-    }
-
-    let textfield;
-    if (this.props.input && this.props.meta) {
-      textfield = (
-        <TextField
-          data-test-timepicker-input
-          inputRef={this.textfieldRef}
-          input={input}
-          value={this.getPresentationValue()}
-          label={label}
-          aria-label={rest['aria-label'] || ariaLabel}
-          endControl={(readOnly || disabled) ? null : endElement}
-          placeholder={` ${this._timeFormat}`}
-          onKeyDown={(readOnly || disabled) ? null : this.handleKeyDown}
-          onClick={(readOnly || disabled) ? null : this.handleFieldClick}
-          onBlur={this.hideOnBlur}
-          meta={this.props.meta}
-          onChange={this.handleFieldChange}
-          readOnly={readOnly}
-          disabled={disabled}
-          id={this.testId}
-          required={required}
-          hasClearIcon={false}
-          autoFocus={this.props.autoFocus}
-          marginBottom0={marginBottom0}
+  const endElements = [
+    <FormattedMessage key="pickButton" id="stripes-components.showOrHideTimepicker">
+      {([showOrHideTimepickerAriaLabel]) => (
+        <IconButton
+          key="timepickerButton"
+          onClick={toggleTimepicker}
+          onKeyDown={handleKeyDown}
+          aria-label={showOrHideTimepickerAriaLabel}
+          id={`timepicker-toggle-button-${testId}`}
+          tabIndex="-1"
+          icon="clock"
         />
-      );
-    } else {
-      textfield = (
-        <TextField
-          data-test-timepicker-input
-          inputRef={this.textfieldRef}
-          label={label}
-          value={this.getPresentationValue()}
-          aria-label={rest['aria-label'] || ariaLabel}
-          endControl={(readOnly || disabled) ? null : endElement}
-          placeholder={this._timeFormat.toUpperCase()}
-          onKeyDown={(readOnly || disabled) ? null : this.handleKeyDown}
-          onBlur={this.hideOnBlur}
-          onClick={(readOnly || disabled) ? null : this.handleFieldClick}
-          onChange={this.handleFieldChange}
-          readOnly={readOnly}
-          disabled={disabled}
-          id={this.testId}
-          required={required}
-          hasClearIcon={false}
-          autoFocus={this.props.autoFocus}
-          marginBottom0={marginBottom0}
-        />
-      );
-    }
+      )}
+    </FormattedMessage>
+  ];
 
-    const renderTextField = (
-      <div style={{ position: 'relative', width: '100%' }} ref={this.containerRef}>
-        <div
-          aria-live="assertive"
-          aria-relevant="additions"
-          className="sr-only"
-          ref={(ref) => { this.srSpace = ref; }}
-        >
-          <div>{this.state.srMessage}</div>
-        </div>
-        {textfield}
+  // if (displayed) {
+  //   endElements.unshift(
+  //     <FormattedMessage key="clearButton" id="stripes-components.clearFieldValue">
+  //       {([clearFieldValueAriaLabel]) => (
+  //         <IconButton
+  //           key="clearButton"
+  //           onClick={handleClearTime}
+  //           aria-label={clearFieldValueAriaLabel}
+  //           id={`timepicker-clear-button-${testId}`}
+  //           tabIndex="-1"
+  //           icon="times-circle-solid"
+  //         />
+  //       )}
+  //     </FormattedMessage>
+  //   );
+  // }
+
+  return (
+    <>
+      <div style={{ position: 'relative', width: '100%' }} ref={container}>
+        <SRStatus ref={srStatus} />
+        <TextField
+          {...inputProps}
+          value={timePair.timeString || ''}
+          inputRef={handleInputRef}
+          endControl={(inputProps.readOnly || inputProps.disabled) ? null : (<>{endElements}</>)}
+          placeholder={timeFormat.toUpperCase()}
+          onChange={handleChange}
+          onClear={handleClearTime}
+        />
+        <input
+          hidden
+          type="text"
+          value={timePair.formatted || ''}
+          onChange={onChangeFormatted}
+          ref={hiddenInput}
+        />
       </div>
-    );
-
-    const renderDropdown = (
-      <RootCloseWrapper onRootClose={this.handleRootClose} ref={this.dropdownRef}>
-        <TimeDropdown
-          onSetTime={this.handlePickTime}
-          onHide={this.hideTimepicker}
-          selectedTime={this.state.timeString}
-          timeFormat={this.state.timeFormat}
-          ref={(ref) => { this.picker = ref; }}
-          rootRef={this.dropdownRef}
-          mainControl={this.textfieldRef.current}
-          onKeyDown={this.handleKeyDown}
-          onBlur={this.hideTimepicker}
-          locale={this.locale}
-          onNavigation={this.handleDateNavigation}
-          id={this.testId}
-          onClose={this.hideTimepicker}
-        />
-      </RootCloseWrapper>
-    );
-
-    const portalElem = usePortal
-      ? document.getElementById('OverlayContainer')
-      : null;
-
-    return (
-      <>
-        {renderTextField}
+      {localeLoaded && (
         <Popper
           placement={placement}
-          isOpen={this.state.showTimepicker}
-          anchorRef={this.containerRef}
-          onToggle={this.toggleTimepicker}
+          isOpen={showDropdown}
+          anchorRef={container}
+          onToggle={toggleTimepicker}
           portal={usePortal && portalElem}
           modifiers={{
             offset: {
@@ -645,14 +445,36 @@ class Timepicker extends React.Component {
             ...modifiers
           }}
         >
-          {renderDropdown}
+          <RootCloseWrapper onRootClose={hideTimepicker} ref={dropdownRef}>
+            <TimeDropdown
+              hoursFormat={timeFormat.includes('A') ? '12' : '24'}
+              onSetTime={handlePickTime}
+              onHide={() => setShowDropdown(false)}
+              selectedTime={timePair.timeString}
+              timeFormat={timeFormat}
+              rootRef={dropdownRef}
+              mainControl={input.current}
+              locale={localeProp || intl.locale}
+              onKeyDown={handleKeyDown}
+              onBlur={() => setShowDropdown(false)}
+              id={testId}
+              onClose={() => setShowDropdown(false)}
+            />
+          </RootCloseWrapper>
         </Popper>
-      </>
-    );
-  }
-}
+      )}
+    </>
+  );
+};
 
 Timepicker.propTypes = propTypes;
-Timepicker.defaultProps = defaultProps;
 
-export default injectIntl(Timepicker);
+export default FormField(
+  Timepicker,
+  ({ meta }) => ({
+    dirty: meta.dirty,
+    error: (meta.touched && meta.error ? meta.error : ''),
+    valid: meta.valid,
+    warning: (meta.touched ? parseMeta(meta, 'warning') : ''),
+  })
+);
