+++ translations/stripes-components/en.json
@@ -39,7 +39,8 @@
   "Timepicker.PeriodLabel": "AM or PM",
   "Timepicker.SetTimeLabel": "Set time",
   "Timepicker.ScreenReaderLabel": "{label} format is {screenReaderFormat} use down arrow to enter timepicker dialog. {screenReaderMessage}",
-  "Timepicker.LeaveDialog": "leave dialog",
+  "Timepicker.formEnd": "End of form",
+  "Timepicker.formStart": "Beginning of form",
   "clearThisField": "Clear this field",
   "validatingInProcess": "Validating...",
   "fieldIsDisabled": "Field is disabled",
